+++ gems/aws-partitions/CHANGELOG.md
@@ -1,10 +1,13 @@
 Unreleased Changes
 ------------------
 
+* Feature - Added `Partition#region?` and `Partition#service?` methods.
+
 1.243.0 (2019-11-21)
 ------------------
 
 * Feature - Added support for enumerating regions for `Aws::ConnectParticipant`.
+
 
 1.242.0 (2019-11-20)
 ------------------
@@ -1349,4 +1352,3 @@ Unreleased Changes
 ------------------
 
 * Feature - Initial release of the `aws-partitions` gem.
-